# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Bringing group travel plans to life",
        "title": "Toggo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/trips/{tripID}/activities": {
            "get": {
                "description": "Retrieves paginated activities for a trip, optionally filtered by category, time of day, and/or date",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activities"
                ],
                "summary": "Get activities by trip",
                "operationId": "getActivitiesByTripID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "tripID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by category name",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by time of day (morning, afternoon, evening)",
                        "name": "time_of_day",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by calendar date (YYYY-MM-DD); activity must have a date range containing this day",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max items per page (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Opaque cursor returned in next_cursor",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ActivityCursorPageResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new activity for a trip",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activities"
                ],
                "summary": "Create activity",
                "operationId": "createActivity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "tripID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create activity request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateActivityRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Activity"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    }
                }
            }
        },
        "/api/v1/trips/{tripID}/activities/parse-link": {
            "post": {
                "description": "Fetches a URL and extracts structured activity fields (name, description, thumbnail) for form autofill. Supports Airbnb, Booking.com, TikTok, Instagram, and generic travel blog URLs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activities"
                ],
                "summary": "Parse link into activity data",
                "operationId": "parseActivityLink",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "tripID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "URL to parse",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ParseLinkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ParsedActivityData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/trips/{tripID}/activities/{activityID}/categories": {
            "get": {
                "description": "Retrieves categories for an activity with pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activities"
                ],
                "summary": "Get activity categories",
                "operationId": "getActivityCategories",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "tripID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Activity ID",
                        "name": "activityID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Max items per page (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Opaque cursor returned in next_cursor",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ActivityCategoriesPageResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errs.APIError": {
            "type": "object",
            "properties": {
                "message": {},
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "models.ActivityCategoriesPageResult": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "next_cursor": {
                    "type": "string"
                }
            }
        },
        "models.ActivityCursorPageResult": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ActivityAPIResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "next_cursor": {
                    "type": "string"
                }
            }
        },
...
        "models.CreateActivityRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "category_names": {
                    "type": "array",
                    "maxItems": 10,
                    "items": {
                        "type": "string"
                    }
                },
                "dates": {
                    "type": "array",
                    "maxItems": 20,
                    "items": {
                        "$ref": "#/definitions/models.DateRange"
                    }
                },
                "description": {
                    "type": "string"
                },
                "estimated_price": {
                    "type": "number",
                    "minimum": 0
                },
                "image_ids": {
                    "type": "array",
                    "maxItems": 5,
                    "items": {
                        "type": "string"
                    }
                },
                "location_lat": {
                    "type": "number",
                    "maximum": 90,
                    "minimum": -90
                },
                "location_lng": {
                    "type": "number",
                    "maximum": 180,
                    "minimum": -180
                },
                "location_name": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 1
                },
                "media_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "time_of_day": {
                    "enum": [
                        "morning",
                        "afternoon",
                        "evening"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ActivityTimeOfDay"
                        }
                    ]
                },
                "trip_id": {
                    "type": "string"
                }
            }
        },
...
        "models.ParseLinkRequest": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
...
    }
}